+ Move setFillStyle and setStrokeStyle to draw module.
